move s3 utils from torch/ to tools
1. Move s3 utils torch/testing/_internal/* to tools/
2. Create tools/test/test_stats to test logics in tools
3. make run_test.py import discovery safe-guard
